pyproject.toml
[project]
name = "sladk-agents"
version = "0.1.0"
description = "AI agents in Slack powered by Google ADK and Bolt for Python"
requires-python = ">=3.10"
dependencies = [
    "slack-sdk>=3.40.0",
    "slack-bolt>=1.27.0",
    "slack-cli-hooks>=0.2.0",
    "google-adk>=1.27.2",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "aiohttp>=3.9.5",
    "pypdf>=6.7.1",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pytest>=9.0.1",
    "ruff>=0.14.3",
]

[tool.ruff]
[tool.ruff.lint]
[tool.ruff.format]

[tool.pytest.ini_options]
testpaths = ["tests"]
log_file = "logs/pytest.log"
log_file_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
